Parcel napi helpers
