+++ crates/node-bindings/src/file_system/file_system_napi.rs
@@ -3,7 +3,7 @@ use std::path::Path;
 use napi::JsObject;
 use parcel::file_system::FileSystem;
 
-use crate::helpers::js_callable::JsCallable;
+use parcel_napi_helpers::js_callable::JsCallable;
 
 // TODO error handling
