+++ crates/node-bindings/src/lib.rs
@@ -5,7 +5,6 @@ use std::alloc::alloc;
 #[cfg(target_arch = "wasm32")]
 use std::alloc::Layout;
 
-mod helpers;
 mod init_sentry;
 
 #[cfg(all(target_os = "macos", not(miri)))]
