+++ crates/node-bindings/src/parcel/parcel.rs
@@ -13,10 +13,10 @@ use parcel::Parcel;
 use parcel::ParcelOptions;
 
 use crate::file_system::FileSystemNapi;
-use crate::helpers::anyhow_napi;
 use crate::parcel::parcel::tracing_setup::{
   setup_tracing, ParcelTracingGuard, ParcelTracingOptions,
 };
+use parcel_napi_helpers::anyhow_napi;
 
 mod tracing_setup;
 
@@ -73,7 +73,7 @@ impl ParcelNapi {
       node_worker_count = parcel_options.threads;
     }
 
-    let rpc_host_nodejs = RpcHostNodejs::new(&env, options.rpc, node_worker_count.clone())?;
+    let rpc_host_nodejs = RpcHostNodejs::new(node_worker_count.clone())?;
     parcel_options.rpc = Some(Arc::new(rpc_host_nodejs));
 
     // Return self
@@ -116,13 +116,5 @@ impl ParcelNapi {
   #[napi]
   pub async fn _testing_temp_fs_is_dir(&self, path: String) -> napi::Result<bool> {
     Ok(self.parcel.fs.is_dir(&PathBuf::from(path)))
-  }
-
-  #[napi]
-  pub async fn _testing_rpc_ping(&self) -> napi::Result<()> {
-    if self.parcel.rpc.as_ref().unwrap().ping().is_err() {
-      return Err(napi::Error::from_reason("Failed to run"));
-    }
-    Ok(())
   }
 }
