+++ crates/parcel_napi_helpers/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+authors = ["Devon Govett <devongovett@gmail.com>"]
+name = "parcel_napi_helpers"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+
+[dependencies]
+serde = "1"
+anyhow = "1"
+napi = { version = "2", features = ["serde-json", "napi4", "napi5", "async"] }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+napi = { version = "2", features = ["serde-json"] }
+
+[dev-dependencies]
+napi = { version = "2", features = ["serde-json", "napi4", "napi5", "async", "noop"]}
+
+[build-dependencies]
+napi-build = "2"
