+++ crates/parcel_napi_helpers/build.rs
@@ -1,0 +1,7 @@
+#[cfg(not(target_arch = "wasm32"))]
+extern crate napi_build;
+
+fn main() {
+  #[cfg(not(target_arch = "wasm32"))]
+  napi_build::setup();
+}
