+++ crates/parcel_napi_helpers/src/anyhow.rs
@@ -1,0 +1,6 @@
+use anyhow;
+
+/// Convert anyhow error to napi error
+pub fn anyhow_napi(value: anyhow::Error) -> napi::Error {
+  napi::Error::from_reason(format!("[napi] {}", value.to_string()))
+}
