+++ crates/parcel_plugin_rpc/src/nodejs/mod.rs
@@ -1,8 +1,6 @@
 mod napi;
-mod rpc_conn_message;
 mod rpc_conn_nodejs;
 mod rpc_conns_nodejs;
-mod rpc_host_message;
 mod rpc_host_nodejs;
 mod worker_init;
