--- crates/parcel_plugin_rpc/src/nodejs/rpc_host_message.rs
@@ -1,15 +1,0 @@
-use std::sync::mpsc::Sender;
-
-pub enum RpcHostMessage {
-  Ping {
-    response: Sender<Result<(), String>>,
-  },
-}
-
-impl RpcHostMessage {
-  pub fn get_id(&self) -> u32 {
-    match self {
-      Self::Ping { response: _ } => 0,
-    }
-  }
-}
