+++ crates/parcel_plugin_rpc/src/rpc_host.rs
@@ -3,13 +3,12 @@ use std::sync::Arc;
 use anyhow;
 
 pub type RpcHostRef = Arc<dyn RpcHost>;
-pub type RpcConnectionRef = Arc<dyn RpcConnection>;
+pub type RpcWorkerRef = Arc<dyn RpcWorker>;
 
 pub trait RpcHost: Send + Sync {
-  fn ping(&self) -> anyhow::Result<()>;
-  fn start(&self) -> anyhow::Result<RpcConnectionRef>;
+  fn start(&self) -> anyhow::Result<RpcWorkerRef>;
 }
 
-pub trait RpcConnection: Send + Sync {
+pub trait RpcWorker: Send + Sync {
   fn ping(&self) -> anyhow::Result<()>;
 }
