+++ packages/core/core/src/parcel-v3/worker/worker.js
@@ -1,4 +1,15 @@
 // @flow
-import {ParcelWorker} from '../ParcelWorker';
+import * as napi from '@parcel/rust';
+import type {ResolverNapi} from '../plugins/Resolver';
 
-new ParcelWorker();
+export class ParcelWorker {
+  #resolvers: Map<string, ResolverNapi>;
+
+  ping() {
+    console.log('Hi');
+  }
+
+  registerResolver() {}
+}
+
+napi.registerWorker(new ParcelWorker());
